+++ packages/deposit-screener/push_docker.sh
@@ -1,0 +1,2 @@
+COMMIT_HASH=$(git rev-parse --short HEAD)
+docker push nocturnelabs/deposit-screener:$COMMIT_HASH