+++ packages/deposit-screener/src/cli/index.ts
@@ -4,10 +4,11 @@ import { program } from "commander";
 import run from "./commands/run";
 import * as dotenv from "dotenv";
 import { setupDefaultInstrumentation } from "@nocturne-xyz/offchain-utils";
+import { ACTOR_NAME } from "../types";
 
 export default async function main(): Promise<void> {
   dotenv.config();
-  setupDefaultInstrumentation("deposit-screener");
+  setupDefaultInstrumentation(ACTOR_NAME);
 
   program
     .name("screener-cli")